pyproject.toml
[tool.poetry]
name = "grpc-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "grpc_service"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pre-commit = "^3.3.2"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
grpcio = "^1.55.0"
grpcio-tools = "^1.55.0"
aio-pika = "^9.0.7"
pydantic = "^1.10.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
